requirements.txt
discord.py==2.0.1
psycopg2-binary==2.9.9  # Для работы с PostgreSQL
python-dotenv==1.0.1
SQLAlchemy==2.0.32
pytest==7.1.2
jsonschema==4.4.0